+++ app.js
@@ -26,7 +26,7 @@ db.once("open", function () {
 var indexRouter = require("./routes/index");
 var usersRouter = require("./routes/users");
 var casRouter = require("./routes/cas");
-var formRouter = require("./routes/api/Form");
+var formRouter = require("./routes/api/courseInstance");
 
 var app = express();
 
@@ -51,7 +51,7 @@ app.use(authentication());
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/cas", casRouter);
-app.use("/api/Form", formRouter);
+app.use("/api/courseInstance", formRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
